lift and shift inline function to inl files
