+++ include/engine/EngineUtilities.inl
@@ -1,0 +1,140 @@
+#pragma once
+#include <filesystem>
+#include <spdlog/spdlog.h>
+
+template <typename T>
+void Engine<T>::transformOutput(std::vector<std::vector<std::vector<T>>> &input, std::vector<std::vector<T>> &output) {
+    if (input.size() != 1) {
+        auto msg = "The feature vector has incorrect dimensions!";
+        spdlog::error(msg);
+        throw std::logic_error(msg);
+    }
+
+    output = std::move(input[0]);
+}
+
+template <typename T> void Engine<T>::transformOutput(std::vector<std::vector<std::vector<T>>> &input, std::vector<T> &output) {
+    if (input.size() != 1 || input[0].size() != 1) {
+        auto msg = "The feature vector has incorrect dimensions!";
+        spdlog::error(msg);
+        throw std::logic_error(msg);
+    }
+
+    output = std::move(input[0][0]);
+}
+
+template <typename T>
+cv::cuda::GpuMat Engine<T>::resizeKeepAspectRatioPadRightBottom(const cv::cuda::GpuMat &input, size_t height, size_t width,
+                                                                const cv::Scalar &bgcolor) {
+    float r = std::min(width / (input.cols * 1.0), height / (input.rows * 1.0));
+    int unpad_w = r * input.cols;
+    int unpad_h = r * input.rows;
+    cv::cuda::GpuMat re(unpad_h, unpad_w, CV_8UC3);
+    cv::cuda::resize(input, re, re.size());
+    cv::cuda::GpuMat out(height, width, CV_8UC3, bgcolor);
+    re.copyTo(out(cv::Rect(0, 0, re.cols, re.rows)));
+    return out;
+}
+
+template <typename T> void Engine<T>::getDeviceNames(std::vector<std::string> &deviceNames) {
+    int numGPUs;
+    cudaGetDeviceCount(&numGPUs);
+
+    for (int device = 0; device < numGPUs; device++) {
+        cudaDeviceProp prop;
+        cudaGetDeviceProperties(&prop, device);
+
+        deviceNames.push_back(std::string(prop.name));
+    }
+}
+
+template <typename T> std::string Engine<T>::serializeEngineOptions(const Options &options, const std::string &onnxModelPath) {
+    const auto filenamePos = onnxModelPath.find_last_of('/') + 1;
+    std::string engineName = onnxModelPath.substr(filenamePos, onnxModelPath.find_last_of('.') - filenamePos) + ".engine";
+
+    // Add the GPU device name to the file to ensure that the model is only used
+    // on devices with the exact same GPU
+    std::vector<std::string> deviceNames;
+    getDeviceNames(deviceNames);
+
+    if (static_cast<size_t>(options.deviceIndex) >= deviceNames.size()) {
+        auto msg = "Error, provided device index is out of range!";
+        spdlog::error(msg);
+        throw std::runtime_error(msg);
+    }
+
+    auto deviceName = deviceNames[options.deviceIndex];
+    // Remove spaces from the device name
+    deviceName.erase(std::remove_if(deviceName.begin(), deviceName.end(), ::isspace), deviceName.end());
+
+    engineName += "." + deviceName;
+
+    // Serialize the specified options into the filename
+    if (options.precision == Precision::FP16) {
+        engineName += ".fp16";
+    } else if (options.precision == Precision::FP32) {
+        engineName += ".fp32";
+    } else {
+        engineName += ".int8";
+    }
+
+    engineName += "." + std::to_string(options.maxBatchSize);
+    engineName += "." + std::to_string(options.optBatchSize);
+
+    spdlog::info("Engine name: {}", engineName);
+    return engineName;
+}
+
+template <typename T>
+cv::cuda::GpuMat Engine<T>::blobFromGpuMats(const std::vector<cv::cuda::GpuMat> &batchInput, const std::array<float, 3> &subVals,
+                                            const std::array<float, 3> &divVals, bool normalize, bool swapRB) {
+   
+    CHECK(!batchInput.empty())
+    CHECK(batchInput[0].channels() == 3)
+    
+    cv::cuda::GpuMat gpu_dst(1, batchInput[0].rows * batchInput[0].cols * batchInput.size(), CV_8UC3);
+
+    size_t width = batchInput[0].cols * batchInput[0].rows;
+    if (swapRB) {
+        for (size_t img = 0; img < batchInput.size(); ++img) {
+            std::vector<cv::cuda::GpuMat> input_channels{
+                cv::cuda::GpuMat(batchInput[0].rows, batchInput[0].cols, CV_8U, &(gpu_dst.ptr()[width * 2 + width * 3 * img])),
+                cv::cuda::GpuMat(batchInput[0].rows, batchInput[0].cols, CV_8U, &(gpu_dst.ptr()[width + width * 3 * img])),
+                cv::cuda::GpuMat(batchInput[0].rows, batchInput[0].cols, CV_8U, &(gpu_dst.ptr()[0 + width * 3 * img]))};
+            cv::cuda::split(batchInput[img], input_channels); // HWC -> CHW
+        }
+    } else {
+        for (size_t img = 0; img < batchInput.size(); ++img) {
+            std::vector<cv::cuda::GpuMat> input_channels{
+                cv::cuda::GpuMat(batchInput[0].rows, batchInput[0].cols, CV_8U, &(gpu_dst.ptr()[0 + width * 3 * img])),
+                cv::cuda::GpuMat(batchInput[0].rows, batchInput[0].cols, CV_8U, &(gpu_dst.ptr()[width + width * 3 * img])),
+                cv::cuda::GpuMat(batchInput[0].rows, batchInput[0].cols, CV_8U, &(gpu_dst.ptr()[width * 2 + width * 3 * img]))};
+            cv::cuda::split(batchInput[img], input_channels); // HWC -> CHW
+        }
+    }
+    cv::cuda::GpuMat mfloat;
+    if (normalize) {
+        // [0.f, 1.f]
+        gpu_dst.convertTo(mfloat, CV_32FC3, 1.f / 255.f);
+    } else {
+        // [0.f, 255.f]
+        gpu_dst.convertTo(mfloat, CV_32FC3);
+    }
+
+    // Apply scaling and mean subtraction
+    cv::cuda::subtract(mfloat, cv::Scalar(subVals[0], subVals[1], subVals[2]), mfloat, cv::noArray(), -1);
+    cv::cuda::divide(mfloat, cv::Scalar(divVals[0], divVals[1], divVals[2]), mfloat, 1, -1);
+
+    return mfloat;
+}
+
+template <typename T> void Engine<T>::clearGpuBuffers() {
+    if (!m_buffers.empty()) {
+        // Free GPU memory of outputs
+        const auto numInputs = m_inputDims.size();
+        for (int32_t outputBinding = numInputs; outputBinding < m_engine->getNbIOTensors(); ++outputBinding) {
+            Util::checkCudaErrorCode(cudaFree(m_buffers[outputBinding]));
+        }
+        m_buffers.clear();
+    }
+}
